move utils to their own folder
