--- utilities/cookie/index.ts
@@ -1,31 +1,0 @@
-// ANCHOR: Set Cookie
-// DOCS: https://www.w3schools.com/js/js_cookies.asp
-export function setCookie(cname, cvalue, exdays) {
-  const d = new Date();
-  d.setTime(d.getTime() + (exdays * 24 * 60 * 60 * 1000));
-  let expires = "expires="+d.toUTCString();
-  document.cookie = cname + "=" + cvalue + ";" + expires + ";path=/";
-};
-
-// ANCHOR: Get Cookie
-// DOCS: https://www.w3schools.com/js/js_cookies.asp
-export function getCookie(cname) {
-  let name = cname + "=";
-  let ca = document.cookie.split(';');
-  for(let i = 0; i < ca.length; i++) {
-    let c = ca[i];
-    while (c.charAt(0) == ' ') {
-      c = c.substring(1);
-    }
-    if (c.indexOf(name) == 0) {
-      return c.substring(name.length, c.length);
-    }
-  }
-  return "";
-};
-
-// ANCHOR: Clear Cookie
-// DOCS: https://www.w3schools.com/js/js_cookies.asp
-export function clearCookie(cname) {
-  setCookie(cname, "", 0);
-};
