+++ utilities/index.ts
@@ -1,0 +1,15 @@
+// ANCHOR: General Utilities
+export { checkWindowObject } from "./modules/checkWindowObject";
+
+// ANCHOR: Cookie Utilities
+export { clearCookie } from "./modules/cookie/clearCookie";
+export { getCookie } from "./modules/cookie/getCookie";
+export { setCookie } from "./modules/cookie/setCookie";
+
+// ANCHOR: Error Utilities
+export { checkUserState } from "./modules/checkUserState";
+export { clearUserToken } from "./modules/clearUserToken";
+export { setUserToken } from "./modules/setUserToken";
+
+// ANCHOR: User Utilities
+export { errorLog } from "./modules/errorLog";
