+++ utilities/modules/checkUserState/index.ts
@@ -3,8 +3,8 @@ import { reduxStore } from '@state/store';
 import { userTokenAction } from '@state/actions';
 
 // ANCHOR: Utilities
-import { getCookie } from "@utilities/cookie";
-import { checkWindowObject } from "@utilities/checkWindowObject";
+import { getCookie } from "@utilities/index";
+import { checkWindowObject } from "@utilities/index";
 
 export function checkUserState() {
   if (checkWindowObject) {
