+++ utilities/modules/checkWindowObject/index.ts
