+++ utilities/modules/cookie/clearCookie.ts
@@ -1,0 +1,8 @@
+// ANCHOR: Cookie Import
+import { setCookie } from "./setCookie";
+
+// ANCHOR: Clear Cookie
+// DOCS: https://www.w3schools.com/js/js_cookies.asp
+export function clearCookie(cname) {
+  setCookie(cname, "", 0);
+};
