+++ utilities/modules/cookie/getCookie.ts
@@ -1,0 +1,16 @@
+// ANCHOR: Get Cookie
+// DOCS: https://www.w3schools.com/js/js_cookies.asp
+export function getCookie(cname) {
+  let name = cname + "=";
+  let ca = document.cookie.split(';');
+  for(let i = 0; i < ca.length; i++) {
+    let c = ca[i];
+    while (c.charAt(0) == ' ') {
+      c = c.substring(1);
+    }
+    if (c.indexOf(name) == 0) {
+      return c.substring(name.length, c.length);
+    }
+  }
+  return "";
+};
