+++ utilities/modules/cookie/setCookie.ts
@@ -1,0 +1,8 @@
+// ANCHOR: Set Cookie
+// DOCS: https://www.w3schools.com/js/js_cookies.asp
+export function setCookie(cname, cvalue, exdays) {
+  const d = new Date();
+  d.setTime(d.getTime() + (exdays * 24 * 60 * 60 * 1000));
+  let expires = "expires="+d.toUTCString();
+  document.cookie = cname + "=" + cvalue + ";" + expires + ";path=/";
+};
