+++ utilities/modules/errorLog/index.ts
